# docs/swagger.yaml
basePath: /api/v1
definitions:
...
  service.AIGenerateRequest:
    properties:
      prompt:
        type: string
    required:
    - prompt
    type: object
  service.AIGenerateResponse:
    properties:
      image_url:
        type: string
      message:
        type: string
    type: object
  service.AddToCartRequest:
    properties:
      color:
        type: string
      design_id:
        type: integer
      product_id:
        type: integer
      quantity:
        type: integer
      size:
        type: string
    required:
    - color
    - design_id
    - product_id
    - quantity
    - size
    type: object
  service.CartDesignResponse:
    properties:
      id:
        type: integer
      image_url:
        type: string
      prompt:
        type: string
      style:
        type: string
    type: object
  service.CartItemResponse:
    properties:
      color:
        type: string
      design:
        $ref: '#/definitions/service.CartDesignResponse'
      id:
        type: integer
      product:
        $ref: '#/definitions/service.ProductResponse'
      quantity:
        type: integer
      size:
        type: string
    type: object
  service.CartResponse:
    properties:
      items:
        items:
          $ref: '#/definitions/service.CartItemResponse'
        type: array
      total_items:
        type: integer
      total_value:
        type: number
    type: object
...
  service.ProductResponse:
    properties:
      base_price:
        type: number
      brand:
        type: string
      category:
        type: string
      created_at:
        type: string
      description:
        type: string
      id:
        type: integer
      is_active:
        type: boolean
      material:
        type: string
      name:
        type: string
    type: object
  service.UpdateCartRequest:
    properties:
      quantity:
        type: integer
    required:
    - quantity
    type: object
host: localhost:8080
info:
  contact:
    email: support@swagger.io
    name: API Support
    url: http://www.swagger.io/support
  description: AI驱动的T恤设计商店API服务，提供用户认证、AI设计生成等功能
  license:
    name: MIT
    url: https://opensource.org/licenses/MIT
  termsOfService: http://swagger.io/terms/
  title: AI T-Shirt Shop API
  version: "1.0"
paths:
...
  /cart:
    get:
      consumes:
      - application/json
      description: 获取当前用户的购物车内容
      produces:
      - application/json
      responses:
        "200":
          description: 购物车内容
          schema:
            $ref: '#/definitions/service.CartResponse'
        "401":
          description: 未授权
          schema:
            additionalProperties: true
            type: object
        "500":
          description: 内部服务器错误
          schema:
            additionalProperties: true
            type: object
      security:
      - BearerAuth: []
      summary: 获取购物车
      tags:
      - cart
  /cart/{id}:
    delete:
      consumes:
      - application/json
      description: 从购物车中删除指定的商品
      parameters:
      - description: 购物车商品ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: 删除成功
          schema:
            additionalProperties: true
            type: object
        "400":
          description: 请求参数错误
          schema:
            additionalProperties: true
            type: object
        "401":
          description: 未授权
          schema:
            additionalProperties: true
            type: object
        "500":
          description: 内部服务器错误
          schema:
            additionalProperties: true
            type: object
      security:
      - BearerAuth: []
      summary: 从购物车删除商品
      tags:
      - cart
    put:
      consumes:
      - application/json
      description: 更新购物车中指定商品的数量
      parameters:
      - description: 购物车商品ID
        in: path
        name: id
        required: true
        type: integer
      - description: 更新购物车请求参数
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/service.UpdateCartRequest'
      produces:
      - application/json
      responses:
        "200":
          description: 更新成功
          schema:
            additionalProperties: true
            type: object
        "400":
          description: 请求参数错误
          schema:
            additionalProperties: true
            type: object
        "401":
          description: 未授权
          schema:
            additionalProperties: true
            type: object
        "404":
          description: 购物车商品不存在
          schema:
            additionalProperties: true
            type: object
        "500":
          description: 内部服务器错误
          schema:
            additionalProperties: true
            type: object
      security:
      - BearerAuth: []
      summary: 更新购物车商品数量
      tags:
      - cart
  /cart/add:
    post:
      consumes:
      - application/json
      description: 将指定的商品和设计组合添加到用户购物车
      parameters:
      - description: 添加到购物车请求参数
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/service.AddToCartRequest'
      produces:
      - application/json
      responses:
        "200":
          description: 添加成功
          schema:
            additionalProperties: true
            type: object
        "400":
          description: 请求参数错误
          schema:
            additionalProperties: true
            type: object
        "401":
          description: 未授权
          schema:
            additionalProperties: true
            type: object
        "500":
          description: 内部服务器错误
          schema:
            additionalProperties: true
            type: object
      security:
      - BearerAuth: []
      summary: 添加商品到购物车
      tags:
      - cart
  /cart/clear:
    delete:
      consumes:
      - application/json
      description: 清空当前用户的整个购物车
      produces:
      - application/json
      responses:
        "200":
          description: 清空成功
          schema:
            additionalProperties: true
            type: object
        "401":
          description: 未授权
          schema:
            additionalProperties: true
            type: object
        "500":
          description: 内部服务器错误
          schema:
            additionalProperties: true
            type: object
      security:
      - BearerAuth: []
      summary: 清空购物车
      tags:
      - cart
  /designs/generate:
    post:
      consumes:
      - application/json
      description: 使用AI根据提示词生成T恤设计图案
      parameters:
      - description: AI生成请求参数
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/service.AIGenerateRequest'
      produces:
      - application/json
      responses:
        "200":
          description: 设计生成成功
          schema:
            $ref: '#/definitions/service.AIGenerateResponse'
        "400":
          description: 请求参数错误
          schema:
            additionalProperties: true
            type: object
        "401":
          description: 未授权
          schema:
            additionalProperties: true
            type: object
        "500":
          description: 内部服务器错误
          schema:
            additionalProperties: true
            type: object
      security:
      - BearerAuth: []
      summary: 生成AI设计
      tags:
      - designs
  /designs/my-designs:
    get:
      consumes:
      - application/json
      description: 获取当前用户的所有设计作品
      produces:
      - application/json
      responses:
        "200":
          description: 用户设计列表
          schema:
            additionalProperties: true
            type: object
        "401":
          description: 未授权
          schema:
            additionalProperties: true
            type: object
        "500":
          description: 内部服务器错误
          schema:
            additionalProperties: true
            type: object
      security:
      - BearerAuth: []
      summary: 获取用户设计
      tags:
      - designs
...
securityDefinitions:
  BearerAuth:
    description: Bearer token for API authentication
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"